refactor(Dashboard):重构个人仪表盘查询和数据结构- 修改 DashboardController 中 GetPersonalDashbboard 方法签名- 更新 DashboardPersonalResponseQuery 数据结构，调整字段类型-优化 DashboardRepository 中的 SQL 查询语句，使用 FILTER语法- 调整 DashboardPersonalResponseQuery与数据库查询结果的映射关系

+++ Dashboard.Domain/Dtos/DashboardPersonalResponseQuery.cs
@@ -1,9 +1,9 @@
 namespace Dashboard.Domain.Dtos;
 
 public record DashboardPersonalResponseQuery(
-    int ConversasResolvidas,
-    int ConversasAtivas,
-    int MediaAvaliacao,
-    int ConversasPendentes,
-    int ConversasEmAndamento
+    long ConversasResolvidas,
+    long ConversasAtivas,
+    decimal MediaAvaliacao,
+    long ConversasPendentes,
+    long ConversasEmAndamento
 );
